move the functions getZendeskSyncWorkflowId and getZendeskGarbageCollectionWorkflowId to types
