+++ types/src/connectors/zendesk.ts
@@ -1,0 +1,11 @@
+import { ModelId } from "../shared/model_id";
+
+export function getZendeskSyncWorkflowId(connectorId: ModelId): string {
+  return `zendesk-sync-${connectorId}`;
+}
+
+export function getZendeskGarbageCollectionWorkflowId(
+  connectorId: ModelId
+): string {
+  return `zendesk-gc-${connectorId}`;
+}
